basic local GET route
