+++ .gitignore
@@ -3,3 +3,5 @@ go.mod
 go.sum
 raw.html
 json/
+response/
+*.exe
