+++ getFunctions/README.md
@@ -1,0 +1,1 @@
+Functions used in API routes to retrieve data
